Update Java version to 20 and refactor Github Workflow
This commit updates the application Java version to 20 in both the Gradle configuration and the CI environment variable. The base Docker image in the Jib configuration has been updated accordingly to use the Temurin JDK 20.

The CI workflow has also been refactored for brevity and readability. The if conditions have been simplified and Gradle caching has been consolidated. The syntax of the Set Gradle permissions action was also updated. These changes are aimed at reducing complexity and improving maintainability.